Fix up uses of sortBy and filterOn
