+++ components/forms/FormSummary.js
@@ -10,15 +10,15 @@ export default class FormSummary extends Component {
       content = (
         <div className="alert alert-danger">
           {errors._.map(error => {
-            const text = error.hasOwnProperty('reason') ? error.reason : error;
+            const text = error.reason || error;
             return (<div key={text}>{text}</div>);
           })}
         </div>
       );
-    } else if (Object.keys(errors).length === 1) {
-      content = success ? <div className="alert alert-success">{success}</div> : '';
-    } else if (errors._) {
+    } else if (Object.keys(errors).length > 0) {
       content = <div className="alert alert-danger">Please fix all errors before retrying.</div>;
+    } else if (errors._) {
+      content = success ? <div className="alert alert-success">{success}</div> : '';
     }
 
     return (
