+++ src/utilities.js
@@ -5,6 +5,7 @@ import { DeleteModalBody } from 'linode-components/modals';
 
 import { showModal, hideModal } from '~/actions/modal';
 import { removeSelected } from '~/actions/select';
+import { dispatchOrStoreErrors } from '~/api/util';
 
 
 export function convertUnits(value, units, unitType, fixedNumber = 0) {
@@ -25,11 +26,13 @@ export function confirmThenDelete(dispatch, objectLabel, deleteFunction, objectT
 
     dispatch(showModal(title, (
       <DeleteModalBody
-        onSubmit={async function() {
+        onSubmit={function() {
           const ids = toDelete.map(o => o[idKey]);
-          await Promise.all(ids.map(id => dispatch(deleteFunction(id))));
-          dispatch(removeSelected(objectType, ids));
-          dispatch(hideModal());
+          return dispatch(dispatchOrStoreErrors.call(this, [
+            () => (dispatch) => Promise.all(ids.map(id => dispatch(deleteFunction(id)))),
+            () => removeSelected(objectType, ids),
+            hideModal,
+          ]));
         }}
         items={toDelete.map(labelFn)}
         typeOfItem={`${objectLabel}s`}
